refactor: move vendor script to tools/vendor/ package
Move scripts/vendor_llama_cpp.py to tools/vendor/llama_cpp.py as a
proper Python package. Clean up test imports (remove sys.path hack),
group tests into pytest classes, remove divider comments, and update
CI workflow to use python -m tools.vendor.llama_cpp.
